move search to gateway, add list helpers for serializing federated vs local data
