+++ gateway/sds_gateway/templates/users/components/dataset_list_modals.html
@@ -4,15 +4,17 @@ Renders the same modal set as dataset_list.html for the current page_obj.
 Context: page_obj (paginated dataset list)
 {% endcomment %}
 {% for dataset in page_obj %}
-    {% include "users/partials/web_download_modal.html" with item=dataset item_type="dataset" %}
-    {% include "users/partials/sdk_download_modal.html" with dataset=dataset %}
-    {% if dataset.is_owner or dataset.permission_level == 'co-owner' or dataset.permission_level == 'contributor' %}
-        {% include "users/partials/share_modal.html" with item=dataset item_type="dataset" %}
-    {% endif %}
-    {% if dataset.is_owner or dataset.permission_level == 'co-owner' %}
-        {% include "users/partials/dataset_version_control.html" with dataset=dataset %}
-        {% if not dataset.dataset.status == 'final' or not dataset.is_public %}
-            {% include "users/partials/publish_dataset_modal.html" with dataset=dataset %}
+    {% if not dataset.is_federated %}
+        {% include "users/partials/web_download_modal.html" with item=dataset item_type="dataset" %}
+        {% include "users/partials/sdk_download_modal.html" with dataset=dataset %}
+        {% if dataset.is_owner or dataset.permission_level == 'co-owner' or dataset.permission_level == 'contributor' %}
+            {% include "users/partials/share_modal.html" with item=dataset item_type="dataset" %}
+        {% endif %}
+        {% if dataset.is_owner or dataset.permission_level == 'co-owner' %}
+            {% include "users/partials/dataset_version_control.html" with dataset=dataset %}
+            {% if not dataset.status == 'final' or not dataset.is_public %}
+                {% include "users/partials/publish_dataset_modal.html" with dataset=dataset %}
+            {% endif %}
         {% endif %}
     {% endif %}
 {% endfor %}
