+++ gateway/sds_gateway/templates/users/components/dataset_list_table_row.html
@@ -29,6 +29,9 @@ Context:
             <span class="ms-2 badge bg-secondary">{{ dataset.status_display }}</span>
         {% endif %}
         {% if dataset.is_public %}<i class="bi bi-globe me-1 text-primary" title="Public"></i>{% endif %}
+        {% if dataset.is_federated %}
+            <span class="ms-2 badge bg-info text-dark" title="Federated site">{{ dataset.site_name }}</span>
+        {% endif %}
         {% if dataset.is_shared_with_me %}
             <span class="ms-2 align-middle"
                   data-bs-toggle="tooltip"
@@ -67,16 +70,20 @@ Context:
         {% endif %}
     </td>
     <td class="text-nowrap">
-        {% if dataset.dataset.created_at %}
+        {% if dataset.created_at %}
             {% localtime on %}
             <div class="d-flex align-items-center gap-2">
-                <span class="bg-transparent pe-2">{{ dataset.dataset.created_at|date:"Y-m-d" }}</span>
-                <span class="text-muted bg-transparent">{{ dataset.dataset.created_at|date:"H:i:s T" }}</span>
+                <span class="bg-transparent pe-2">{{ dataset.created_at|date:"Y-m-d" }}</span>
+                <span class="text-muted bg-transparent">{{ dataset.created_at|date:"H:i:s T" }}</span>
             </div>
         {% endlocaltime %}
     {% else %}
         -
     {% endif %}
 </td>
-<td class="text-center">{% include "users/components/dropdown_menu.html" with items=dropdown_menu_items %}</td>
+<td class="text-center">
+    {% if dropdown_menu_items %}
+        {% include "users/components/dropdown_menu.html" with items=dropdown_menu_items %}
+    {% endif %}
+</td>
 </tr>
